Extract interface declarations
