+++ mirai-core-api/src/commonMain/kotlin/contact/file/AbsoluteFile.kt
@@ -1,0 +1,96 @@
+/*
+ * Copyright 2019-2021 Mamoe Technologies and contributors.
+ *
+ * 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
+ * Use of this source code is governed by the GNU AGPLv3 license that can be found through the following link.
+ *
+ * https://github.com/mamoe/mirai/blob/dev/LICENSE
+ */
+
+@file:JvmBlockingBridge
+@file:Suppress("OVERLOADS_INTERFACE")
+
+package net.mamoe.mirai.contact.file
+
+import kotlinx.io.errors.IOException
+import net.mamoe.kjbb.JvmBlockingBridge
+import net.mamoe.mirai.contact.PermissionDeniedException
+import net.mamoe.mirai.message.data.FileMessage
+import net.mamoe.mirai.utils.ExternalResource
+import net.mamoe.mirai.utils.NotStableForInheritance
+
+/**
+ * 绝对文件标识. 精确表示一个远程文件. 不会受同名文件或目录的影响.
+ *
+ * @since 2.8
+ */
+@NotStableForInheritance
+public interface AbsoluteFile : AbsoluteFileFolder {
+    override val isFile: Boolean get() = true
+    override val isFolder: Boolean get() = false
+
+    /**
+     * 文件到期时间, UTC+8 时间戳秒.
+     */
+    public val expiryTime: Long
+
+    /**
+     * 上传者 ID.
+     */
+    public val uploaderId: Long
+
+    /**
+     * 文件大小 (占用空间) bytes.
+     */
+    public val size: Long
+
+    /**
+     * 文件内容 SHA-1.
+     */
+    public val sha1: ByteArray
+
+    /**
+     * 文件内容 MD5.
+     */
+    public val md5: ByteArray
+
+    /**
+     * 移动远程文件到 [folder] 目录下. 成功时返回 `true`, 当远程文件不存在时返回 `false`.
+     *
+     * 注意该操作有可能产生同名文件或目录 (当 [folder] 中已经存在一个名称为 [name] 的文件或目录时).
+     *
+     * @throws IOException 当发生网络错误时可能抛出
+     * @throws IllegalStateException 当发生已知的协议错误时抛出
+     * @throws PermissionDeniedException 当无管理员权限时抛出 (若群仅允许管理员上传)
+     */
+    public suspend fun moveTo(folder: AbsoluteFolder): Boolean
+
+    /**
+     * 覆盖远程文件内容为 [content]. 当远程文件不存在时返回 `false`.
+     *
+     * @param content 新文件内容
+     *
+     * @throws IOException 当发生网络错误时可能抛出
+     * @throws IllegalStateException 当发生已知的协议错误时抛出
+     * @throws PermissionDeniedException 当无管理员权限时抛出 (若群仅允许管理员上传)
+     */
+    @JvmOverloads
+    public suspend fun overrideWith(content: ExternalResource): Boolean
+
+    /**
+     * 将在线文件转换为 [ExternalResource]. 注意该函数不会立即下载文件. 若在之后的使用中远程文件被删除, 则无法确定 [ExternalResource.inputStream] 是否还可用 (不可用时会抛出异常).
+     */
+    public suspend fun asResource(): ExternalResource
+
+    /**
+     * 获得下载链接 URL 字符串.
+     */
+    public suspend fun getUrl(): String
+
+    /**
+     * 得到表示远程文件的可以发送的 [FileMessage].
+     *
+     * 在 [上传文件][RemoteFiles.uploadNewFile] 时就已经发送了文件消息. [toMessage] 可供之后再次发送使用.
+     */
+    public fun toMessage(): FileMessage
+}
