+++ mirai-core-api/src/commonMain/kotlin/contact/file/RemoteFiles.kt
@@ -13,11 +13,9 @@
 package net.mamoe.mirai.contact.file
 
 import kotlinx.coroutines.flow.Flow
-import kotlinx.io.errors.IOException
 import net.mamoe.kjbb.JvmBlockingBridge
 import net.mamoe.mirai.contact.FileSupported
 import net.mamoe.mirai.contact.PermissionDeniedException
-import net.mamoe.mirai.message.data.FileMessage
 import net.mamoe.mirai.utils.ExternalResource
 import net.mamoe.mirai.utils.JavaFriendlyAPI
 import net.mamoe.mirai.utils.NotStableForInheritance
@@ -129,317 +127,3 @@ public interface RemoteFiles {
     ): AbsoluteFile
 }
 
-/**
- * 绝对文件或目录标识. 精确表示一个远程文件. 不会受同名文件或目录的影响.
- *
- * @since 2.8
- */
-@NotStableForInheritance
-public sealed interface AbsoluteFileFolder {
-    /**
-     * 该对象所属 [FileSupported]
-     */
-    public val contact: FileSupported
-
-    /**
-     * 上级 [AbsoluteFileFolder].
-     *
-     * - 当该 [AbsoluteFileFolder] 表示一个目录中的文件时返回文件所属目录的 [AbsoluteFolder].
-     * - 当该 [AbsoluteFileFolder] 表示子目录时返回父目录的 [AbsoluteFolder].
-     *
-     * 特别地,
-     * - 当该 [AbsoluteFileFolder] 表示根目录下的一个文件时返回根目录的 [AbsoluteFolder].
-     * - 当该 [AbsoluteFileFolder] 表示根目录时返回 `null` (表示无上级).
-     *
-     * 也就是说, 若 [AbsoluteFileFolder.parent] 为 `null`, 那么该 [AbsoluteFileFolder] 就表示根目录.
-     */
-    public val parent: AbsoluteFolder?
-
-    /**
-     * 文件或目录的 ID, 即 `fileId` 和 `folderId`. 该属性由服务器维护, 通常唯一且持久.
-     */
-    public val id: String
-
-    /**
-     * 文件名或目录名. 该属性不会因 [renameTo] 等操作而变化.
-     *
-     * 不能包含 `:*?"<>|/\` 任一字符.
-     */
-    public val name: String
-
-    /**
-     * 绝对路径, 如 `/foo/bar.txt`
-     */
-    public val absolutePath: String
-
-    /**
-     * 表示远程文件时返回 `true`.
-     */
-    public val isFile: Boolean
-
-    /**
-     * 表示远程目录时返回 `true`.
-     */
-    public val isFolder: Boolean
-
-    /**
-     * 远程文件或目录的创建时间, UTC+8 时间戳秒.
-     */
-    public val creationTime: Long
-
-    /**
-     * 远程文件或目录的最后修改时间, UTC+8 时间戳秒.
-     */
-    public val lastModifiedTime: Long
-
-
-    /**
-     * 重命名远程文件或目录, **并且**修改当前(`this`) [AbsoluteFileFolder] 的 [name]. 成功时返回 `true`, 当远程文件或目录不存在时返回 `false`.
-     *
-     * 注意该操作有可能产生同名文件或目录 (当服务器已经存在一个名称为 [newName] 的文件或目录时).
-     *
-     * @throws IOException 当发生网络错误时可能抛出
-     * @throws IllegalStateException 当发生已知的协议错误时抛出
-     * @throws PermissionDeniedException 当无管理员权限时抛出 (若群仅允许管理员上传)
-     */
-    public suspend fun renameTo(newName: String): Boolean
-
-    /**
-     * 删除远程文件或目录. 只会根据 [id] 精确地删除一个文件或目录, 不会删除其他同名文件或目录.
-     *
-     * 若目录非空, 则会删除目录中的所有文件. 操作目录或非 Bot 自己上传的文件时需要管理员权限, 无管理员权限时抛出异常.
-     *
-     * @throws IOException 当发生网络错误时可能抛出
-     * @throws IllegalStateException 当发生已知的协议错误时抛出
-     * @throws PermissionDeniedException 当无管理员权限时抛出
-     */
-    public suspend fun delete(): Boolean
-
-    /**
-     * 更新当前 [AbsoluteFileFolder] 对象的文件或目录信息 ([lastModifiedTime] 等).
-     *
-     * 该函数会遍历上级目录的所有文件并匹配当前文件, 因此可能会非常慢, 请不要频繁使用.
-     */
-    public suspend fun refresh()
-
-    /**
-     * 返回更新了文件或目录信息 ([lastModifiedTime] 等) 的, 指向相同文件的 [AbsoluteFileFolder]. 不会更新当前 [AbsoluteFileFolder] 对象.
-     *
-     * 该函数会遍历上级目录的所有文件并匹配当前文件, 因此可能会非常慢, 请不要频繁使用.
-     */
-    public suspend fun refreshed(): AbsoluteFileFolder
-}
-
-/**
- * 绝对目录标识. 精确表示一个远程目录. 不会受同名文件或目录的影响.
- *
- * @since 2.8
- */
-@NotStableForInheritance
-public interface AbsoluteFolder : AbsoluteFileFolder {
-    override val isFile: Boolean get() = false
-    override val isFolder: Boolean get() = true
-
-    /**
-     * 目录内文件数量. 若该目录表示根目录, [contentsCount] 返回 `0`. (无法快速获取)
-     */
-    public val contentsCount: Long
-
-    /**
-     * 当该目录为空时返回 `true`.
-     */
-    public suspend fun isEmpty(): Boolean
-
-
-    ///////////////////////////////////////////////////////////////////////////
-    // list children
-    ///////////////////////////////////////////////////////////////////////////
-
-    /**
-     * 获取该目录下所有子目录列表.
-     */
-    public suspend fun folders(): Flow<AbsoluteFolder>
-
-    /**
-     * 获取该目录下所有子目录列表.
-     *
-     * 实现细节: 为了适合 Java 调用, 实现类似为阻塞式的 [folders], 因此不建议在 Kotlin 使用. 在 Kotlin 请使用 [folders].
-     */
-    @JavaFriendlyAPI
-    public suspend fun foldersStream(): Stream<AbsoluteFolder>
-
-
-    /**
-     * 获取该目录下所有文件列表.
-     */
-    public suspend fun files(): Flow<AbsoluteFile>
-
-    /**
-     * 获取该目录下所有文件列表.
-     *
-     * 实现细节: 为了适合 Java 调用, 实现类似为阻塞式的 [files], 因此不建议在 Kotlin 使用. 在 Kotlin 请使用 [files].
-     */
-    @JavaFriendlyAPI
-    public suspend fun filesStream(): Stream<AbsoluteFile>
-
-
-    /**
-     * 获取该目录下所有文件和子目录列表.
-     */
-    public suspend fun children(): Flow<AbsoluteFileFolder>
-
-    /**
-     * 获取该目录下所有文件和子目录列表.
-     *
-     * 实现细节: 为了适合 Java 调用, 实现类似为阻塞式的 [children], 因此不建议在 Kotlin 使用. 在 Kotlin 请使用 [children].
-     */
-    @JavaFriendlyAPI
-    public suspend fun childrenStream(): Stream<AbsoluteFileFolder>
-
-    ///////////////////////////////////////////////////////////////////////////
-    // resolve and upload
-    ///////////////////////////////////////////////////////////////////////////
-
-    /**
-     * 创建一个名称为 [name] 的子目录. 返回成功创建的或已有的子目录.
-     */
-    public suspend fun createFolder(mame: String): AbsoluteFolder
-
-    /**
-     * 获取一个已存在的名称为 [name] 的子目录. 当该名称的子目录不存在时返回 `null`.
-     */
-    public suspend fun resolveFolder(name: String): AbsoluteFolder?
-
-    /**
-     * 根据路径获取指向的所有名称为 [name] 的文件列表.
-     */
-    public suspend fun resolveFiles(
-        name: String
-    ): Flow<AbsoluteFile>
-
-    /**
-     * 根据路径获取指向的所有名称为 [name] 的文件列表.
-     *
-     * 实现细节: 为了适合 Java 调用, 实现类似为阻塞式的 [resolveFiles], 因此不建议在 Kotlin 使用. 在 Kotlin 请使用 [resolveFiles].
-     */
-    public suspend fun resolveFilesStream(
-        name: String
-    ): Stream<AbsoluteFile>
-
-    /**
-     * 根据路径获取指向的所有名称为 [name] 的文件和目录列表.
-     */
-    public suspend fun resolveAll(
-        name: String
-    ): Flow<AbsoluteFileFolder>
-
-    /**
-     * 根据路径获取指向的所有名称为 [name] 的文件和目录列表.
-     *
-     * 实现细节: 为了适合 Java 调用, 实现类似为阻塞式的 [resolveAll], 因此不建议在 Kotlin 使用. 在 Kotlin 请使用 [resolveAll].
-     */
-    @JavaFriendlyAPI
-    public suspend fun resolveAllStream(
-        name: String
-    ): Stream<AbsoluteFileFolder>
-
-    /**
-     * 上传一个文件到该目录, 返回上传成功的文件标识.
-     *
-     * @param filename 目标文件名
-     * @param content 文件内容
-     * @param keepExisting 为 `false` 时删除远程目录内所有同名文件 (**谨慎操作**, 有可能意想不到地删除多个同名文件), 为 `true` 时则不作操作 (将有可能产生同名文件).
-     *
-     * @throws PermissionDeniedException 当无管理员权限时抛出 (若群仅允许管理员上传)
-     */
-    @JvmOverloads
-    public suspend fun uploadNewFile(
-        filename: String,
-        content: ExternalResource,
-        keepExisting: Boolean = true,
-    ): AbsoluteFile
-
-    public companion object {
-        /**
-         * 根目录 folder ID.
-         * @see id
-         */
-        public const val ROOT_FOLDER_ID: String = "/"
-    }
-}
-
-/**
- * 绝对文件标识. 精确表示一个远程文件. 不会受同名文件或目录的影响.
- *
- * @since 2.8
- */
-@NotStableForInheritance
-public interface AbsoluteFile : AbsoluteFileFolder {
-    override val isFile: Boolean get() = true
-    override val isFolder: Boolean get() = false
-
-    /**
-     * 文件到期时间, UTC+8 时间戳秒.
-     */
-    public val expiryTime: Long
-
-    /**
-     * 上传者 ID.
-     */
-    public val uploaderId: Long
-
-    /**
-     * 文件大小 (占用空间) bytes.
-     */
-    public val size: Long
-
-    /**
-     * 文件内容 SHA-1.
-     */
-    public val sha1: ByteArray
-
-    /**
-     * 文件内容 MD5.
-     */
-    public val md5: ByteArray
-
-    /**
-     * 移动远程文件到 [folder] 目录下. 成功时返回 `true`, 当远程文件不存在时返回 `false`.
-     *
-     * 注意该操作有可能产生同名文件或目录 (当 [folder] 中已经存在一个名称为 [name] 的文件或目录时).
-     *
-     * @throws IOException 当发生网络错误时可能抛出
-     * @throws IllegalStateException 当发生已知的协议错误时抛出
-     * @throws PermissionDeniedException 当无管理员权限时抛出 (若群仅允许管理员上传)
-     */
-    public suspend fun moveTo(folder: AbsoluteFolder): Boolean
-
-    /**
-     * 覆盖远程文件内容为 [content]. 当远程文件不存在时返回 `false`.
-     *
-     * @param content 新文件内容
-     *
-     * @throws IOException 当发生网络错误时可能抛出
-     * @throws IllegalStateException 当发生已知的协议错误时抛出
-     * @throws PermissionDeniedException 当无管理员权限时抛出 (若群仅允许管理员上传)
-     */
-    @JvmOverloads
-    public suspend fun overrideWith(content: ExternalResource): Boolean
-
-    /**
-     * 将在线文件转换为 [ExternalResource]. 注意该函数不会立即下载文件. 若在之后的使用中远程文件被删除, 则无法确定 [ExternalResource.inputStream] 是否还可用 (不可用时会抛出异常).
-     */
-    public suspend fun asResource(): ExternalResource
-
-    /**
-     * 获得下载链接 URL 字符串.
-     */
-    public suspend fun getUrl(): String
-
-    /**
-     * 得到表示远程文件的可以发送的 [FileMessage].
-     *
-     * 在 [上传文件][RemoteFiles.uploadNewFile] 时就已经发送了文件消息. [toMessage] 可供之后再次发送使用.
-     */
-    public fun toMessage(): FileMessage
-}
